[Decomp][Player] Cohesion split: extract item query methods to PlayerItemQuery.cpp
